AdvancedPS

AdvancedPS.PG — Method

PG(n, [resampler = ResampleWithESSThreshold()])
PG(n, [resampler = resample_systematic, ]threshold)

Create a Particle Gibbs sampler with n particles.

If the algorithm for the resampling step is not specified explicitly, systematic resampling is performed if the estimated effective sample size per particle drops below 0.5.

AdvancedPS.ParticleContainer — Type

Data structure for particle filters

effectiveSampleSize(pc :: ParticleContainer)

normalise!(pc::ParticleContainer)

consume(pc::ParticleContainer): return incremental likelihood

AdvancedPS.SMC — Method

SMC(n[, resampler = ResampleWithESSThreshold()])
SMC(n, [resampler = resample_systematic, ]threshold)

Create a sequential Monte Carlo (SMC) sampler with n particles.

If the algorithm for the resampling step is not specified explicitly, systematic resampling is performed if the estimated effective sample size per particle drops below 0.5.

AdvancedPS.TracedRNG — Type

TracedRNG{R,N,T}

Wrapped random number generator from Random123 to keep track of random streams during model evaluation

AdvancedPS.TracedRNG — Type

TracedRNG(r::Random123.AbstractR123=AdvancedPS._BASE_RNG())

Create a TracedRNG with r as the inner RNG.

AdvancedPS.effectiveSampleSize — Method

effectiveSampleSize(pc::ParticleContainer)

Compute the effective sample size $1 / ∑ wᵢ²$, where $wᵢ$`are the normalized weights.

AdvancedPS.getweight — Method

getweight(pc::ParticleContainer, i)

Compute the normalized weight of the ith particle.

AdvancedPS.getweights — Method

getweights(pc::ParticleContainer)

Compute the normalized weights of the particles.

AdvancedPS.inc_counter! — Function

inc_counter!(r::TracedRNG, n::Integer=1)

Increase the model step counter by n

AdvancedPS.increase_logweight! — Method

increase_logweight!(pc::ParticleContainer, i::Int, x)

Increase the unnormalized logarithmic weight of the ith particle with x.

AdvancedPS.load_state! — Method

load_state!(r::TracedRNG)

Load state from current model iteration. Random streams are now replayed

AdvancedPS.logZ — Method

logZ(pc::ParticleContainer)

Return the logarithm of the normalizing constant of the unnormalized logarithmic weights.

AdvancedPS.observe — Method

observe(dist::Distribution, x)

Observe sample x from distribution dist and yield its log-likelihood value.

AdvancedPS.resample_propagate! — Function

resample_propagate!(rng, pc::ParticleContainer[, randcat = resample_systematic,
                    ref = nothing; weights = getweights(pc)])

Resample and propagate the particles in pc.

Function randcat is used for sampling ancestor indices from the categorical distribution of the particle weights. For Particle Gibbs sampling, one can provide a reference particle ref that is ensured to survive the resampling step.

AdvancedPS.resample_stratified — Function

resample_stratified(rng, weights, n)

Return a vector of n samples x₁, ..., xₙ from the numbers 1, ..., length(weights), generated by stratified resampling.

In stratified resampling n ordered random numbers u₁, ..., uₙ are generated, where $uₖ \sim U[(k - 1) / n, k / n)$. Based on these numbers the samples x₁, ..., xₙ are selected according to the multinomial distribution defined by the normalized weights, i.e., xᵢ = j if and only if $uᵢ \in [\sum_{s=1}^{j-1} weights_{s}, \sum_{s=1}^{j} weights_{s})$.

AdvancedPS.resample_systematic — Function

resample_systematic(rng, weights, n)

Return a vector of n samples x₁, ..., xₙ from the numbers 1, ..., length(weights), generated by systematic resampling.

In systematic resampling a random number $u \sim U[0, 1)$ is used to generate n ordered numbers u₁, ..., uₙ where $uₖ = (u + k − 1) / n$. Based on these numbers the samples x₁, ..., xₙ are selected according to the multinomial distribution defined by the normalized weights, i.e., xᵢ = j if and only if $uᵢ \in [\sum_{s=1}^{j-1} weights_{s}, \sum_{s=1}^{j} weights_{s})$.

AdvancedPS.reset_logweights! — Method

reset_logweights!(pc::ParticleContainer)

Reset all unnormalized logarithmic weights to zero.

AdvancedPS.reweight! — Function

reweight!(pc::ParticleContainer)

Check if the final time step is reached, and otherwise reweight the particles by considering the next observation.

AdvancedPS.save_state! — Method

save_state!(r::TracedRNG)

Add current key of the inner rng in r to keys.

AdvancedPS.split — Function

split(key::Integer, n::Integer=1)

Split key into n new keys

AdvancedPS.sweep! — Function

sweep!(rng, pc::ParticleContainer, resampler)

Perform a particle sweep and return an unbiased estimate of the log evidence.

The resampling steps use the given resampler.

Reference

Del Moral, P., Doucet, A., & Jasra, A. (2006). Sequential monte carlo samplers. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 68(3), 411-436.

AdvancedPS.update_keys! — Function

update_keys!(pc::ParticleContainer)

Create new unique keys for the particles in the ParticleContainer

Random.seed! — Method

update_rng!(rng::TracedRNG)

Set key and counter of inner rng in rng to key and the running model step to 0

Random123.set_counter! — Method

set_counter!(r::TracedRNG, n::Integer)

Set the counter of the inner rng in r, used to keep track of the current model step